Move ExtensionMixin from widgets to main package (extensions.ExtensionMixin)
